+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Replaced println with a proper logging
+- Active media players list now updates automatically and immediately
 
 
 ## v0.2.0 - 2022-08-18